flask/requirements copy.txt
Flask==2.0.2
gevent_socketio==0.3.6
Flask_APScheduler==1.12.3
Flask_Bcrypt==1.0.1
Flask_Discord==0.1.64
Flask_Discord_Interactions==1.2.1
Flask_Login==0.5.0
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_Script==2.0.6
Flask_SocketIO==5.1.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
itsdangerous==2.0.1
Pillow==9.0.0
pytest==6.2.5
requests==2.26.0
selenium==4.1.0
SQLAlchemy==1.4.29
typing_extensions==3.10.0.0
WTForms==3.0.1
APScheduler==3.8.1
beautifulsoup4==4.10.0
discord.py==1.7.3
eventlet==0.33.0
email-validator==1.1.3
